+++ tf-modules/archive/reconciliation_report.tf
@@ -15,8 +15,6 @@ resource "aws_lambda_function" "create_reconciliation_report" {
       CMR_HOST                         = var.cmr_custom_host
       DISTRIBUTION_ENDPOINT            = var.distribution_url
       ES_HOST                          = var.elasticsearch_hostname
-      ES_SCROLL                        = lookup(var.elasticsearch_client_config, "create_reconciliation_report_es_scroll_duration", "6m")
-      ES_SCROLL_SIZE                   = lookup(var.elasticsearch_client_config, "create_reconciliation_report_es_scroll_size", 1000)
       stackName                        = var.prefix
       system_bucket                    = var.system_bucket
       cmr_client_id                    = var.cmr_client_id